+++ client/components/blocks/index.js
@@ -1,1 +1,2 @@
+export DiffBlock from './DiffBlock.jsx';
 export HeaderBlock from './HeaderBlock.jsx';